requirements-advanced.txt

# الميزات المتقدمة للمساعد الذكي
tensorflow==2.15.0
pytorch-lightning==2.1.2
huggingface-hub==0.19.4
sentence-transformers==2.2.2

# التحليلات المتقدمة
plotly==5.17.0
dash==2.16.1
matplotlib==3.8.2
seaborn==0.13.0

# قواعد البيانات المتقدمة
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# الأمان والتشفير
cryptography==41.0.8
bcrypt==4.1.2
pyjwt==2.8.0

# التكامل مع الخدمات الخارجية
openai==1.3.8
anthropic==0.8.1
google-cloud-speech==2.24.1
boto3==1.34.0

# أدوات التطوير والاختبار
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
